// src/enums/projects.enum.ts
export enum EProjectPotency {
  MINOR = 'Minore',
  MEDIUM = 'Media',
  MAJOR = 'Maggiore',
  EXTREME = 'Estrema',
}

export enum EProjectArea {
  SELF = 'Individuale',
  SMALL = 'Piccola',
  BIG = 'Grande',
  HUGE = 'Enorme',
}

export enum EProjectEffect {
  NONE = 'Nessuno',
  MINOR = 'Minore',
  HEAVY = 'Pesante',
  MASSIVE = 'Massiccio',
}

export enum EProjectUseType {
  CONSUMABLE = 'Consumabile',
  PERMANENT = 'Permanente',
}

export enum EProjectTinkererLevel {
  LOW = '5+',
  MEDIUM = '20+',
  HIGH = '40+',
}